requirements.txt
amqp==2.4.2
billiard==3.5.0.5
celery==4.2.1
Django==2.1.7
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
kombu==4.4.0
pytz==2018.9
redis==3.2.0
vine==1.2.0
